requirements.txt
Flask==1.0.2
google-cloud-datastore
google-cloud-storage
google-cloud-speech
pillow==4.3.0